+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ## [Unreleased]
 
+### Added
+
+- `Equinox.Core.TaskCellDictionary`: Helper to manage concurrent reads in View processing [#477](https://github.com/jet/equinox/pull/477)
+
 ### Fixed
 
 - `Tool`: Corrected `NullReferenceException` when using `StreamLike` option